Refactor and move common add or update resources on the claim set to separate class.
