+++ Application/EdFi.Ods.AdminApp.Management/ClaimSetEditor/ClaimSetFileImportCommand.cs
@@ -5,7 +5,6 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using EdFi.Ods.AdminApp.Management.ClaimSetEditor.Extensions;
 using EdFi.Ods.AdminApp.Management.Database.Queries;
 
 namespace EdFi.Ods.AdminApp.Management.ClaimSetEditor
@@ -13,22 +12,16 @@ namespace EdFi.Ods.AdminApp.Management.ClaimSetEditor
     public class ClaimSetFileImportCommand
     {
         private readonly AddClaimSetCommand _addClaimSetCommand;
-        private readonly EditResourceOnClaimSetCommand _editResourceOnClaimSetCommand;
         private readonly GetResourceClaimsQuery _getResourceClaimsQuery;
-        private readonly OverrideDefaultAuthorizationStrategyCommand _overrideDefaultAuthorizationStrategyCommand;
-        private readonly IGetClaimSetByNameQuery _claimSetByNameQuery;
+        private readonly AddOrUpdateResourcesOnClaimSetCommand _addOrUpdateResourcesOnClaimSetCommand;
 
         public ClaimSetFileImportCommand(AddClaimSetCommand addClaimSetCommand,
-            EditResourceOnClaimSetCommand editResourceOnClaimSetCommand,
             GetResourceClaimsQuery getResourceClaimsQuery,
-            OverrideDefaultAuthorizationStrategyCommand overrideDefaultAuthorizationStrategyCommand,
-            IGetClaimSetByNameQuery claimSetByNameQuery)
+            AddOrUpdateResourcesOnClaimSetCommand addOrUpdateResourcesOnClaimSetCommand)
         {
             _addClaimSetCommand = addClaimSetCommand;
-            _editResourceOnClaimSetCommand = editResourceOnClaimSetCommand;
             _getResourceClaimsQuery = getResourceClaimsQuery;
-            _overrideDefaultAuthorizationStrategyCommand = overrideDefaultAuthorizationStrategyCommand;
-            _claimSetByNameQuery = claimSetByNameQuery;
+            _addOrUpdateResourcesOnClaimSetCommand = addOrUpdateResourcesOnClaimSetCommand;
         }
 
         public void Execute(SharingModel model)
@@ -37,56 +30,12 @@ namespace EdFi.Ods.AdminApp.Management.ClaimSetEditor
             var allResources = GetDbResources();
             foreach (var claimSet in sharingClaimSets)
             {
-                var claimSetId = AddOrGetClaimSet(claimSet.Name);
-
+                var claimSetId = _addClaimSetCommand.Execute(new AddClaimSetModel
+                {
+                    ClaimSetName = claimSet.Name
+                });
                 var resources = claimSet.ResourceClaims.Select(x => x.ToObject<ResourceClaim>()).ToList();
-                var childResources = new List<ResourceClaim>();
-                foreach (var resourceClaims in resources.Select(x => x.Children))
-                    childResources.AddRange(resourceClaims);
-                resources.AddRange(childResources);
-                var currentResources = resources.Select(r =>
-                    {
-                        var resource = allResources.FirstOrDefault(dr => dr.Name.Equals(r.Name));
-                        if (resource != null)
-                        {
-                            resource.Create = r.Create;
-                            resource.Read = r.Read;
-                            resource.Update = r.Update;
-                            resource.Delete = r.Delete;
-                            resource.AuthStrategyOverridesForCRUD = r.AuthStrategyOverridesForCRUD;
-                        }
-                        return resource;
-                    }).ToList();
-                currentResources.RemoveAll(x => x == null);
-                foreach (var resource in currentResources)
-                {
-                    var editResourceModel = new EditResourceOnClaimSetModel
-                    {
-                        ClaimSetId = claimSetId,
-                        ResourceClaim = resource 
-                    };
-
-                    _editResourceOnClaimSetCommand.Execute(editResourceModel);
-
-                    if (resource.AuthStrategyOverridesForCRUD != null && resource.AuthStrategyOverridesForCRUD.Any())
-                    {
-                        var overrideAuthStrategyModel = new OverrideAuthorizationStrategyModel
-                        {
-                            ClaimSetId = claimSetId,
-                            ResourceClaimId = resource.Id,
-                            AuthorizationStrategyForCreate = AuthStrategyOverrideForAction(resource.AuthStrategyOverridesForCRUD.Create()),
-                            AuthorizationStrategyForRead = AuthStrategyOverrideForAction(resource.AuthStrategyOverridesForCRUD.Read()),
-                            AuthorizationStrategyForUpdate = AuthStrategyOverrideForAction(resource.AuthStrategyOverridesForCRUD.Update()),
-                            AuthorizationStrategyForDelete = AuthStrategyOverrideForAction(resource.AuthStrategyOverridesForCRUD.Delete())
-                        };
-                        _overrideDefaultAuthorizationStrategyCommand.Execute(overrideAuthStrategyModel);
-                    }
-                }
-
-                static int AuthStrategyOverrideForAction(AuthorizationStrategy authorizationStrategy)
-                {
-                    return authorizationStrategy != null ? authorizationStrategy.AuthStrategyId : 0;
-                }
+                _addOrUpdateResourcesOnClaimSetCommand.Execute(claimSetId, resources, allResources);
             }
         }
 
@@ -102,41 +51,5 @@ namespace EdFi.Ods.AdminApp.Management.ClaimSetEditor
 
             return allResources;
         }
-
-        private int AddOrGetClaimSet(string claimSetName)
-        {
-            //var existingClaimSet = _claimSetByNameQuery.Execute(claimSetName);
-
-            //if (existingClaimSet != null)
-            //    return existingClaimSet.Id;
-
-            var claimSetId = _addClaimSetCommand.Execute(new AddClaimSetModel
-            {
-                ClaimSetName = claimSetName
-            });
-            return claimSetId;
-        }
-
-    }
-
-    public class AddClaimSetModel: IAddClaimSetModel
-    {
-        public string ClaimSetName { get; set; }
-    }
-
-    public class EditResourceOnClaimSetModel : IEditResourceOnClaimSetModel
-    {
-        public int ClaimSetId { get; set; }
-        public ResourceClaim ResourceClaim { get; set;  }
-    }
-
-    public class OverrideAuthorizationStrategyModel : IOverrideDefaultAuthorizationStrategyModel
-    {
-        public int ClaimSetId { get; set; }
-        public int ResourceClaimId { get; set; }
-        public int AuthorizationStrategyForCreate { get; set; }
-        public int AuthorizationStrategyForRead { get; set; }
-        public int AuthorizationStrategyForUpdate { get; set; }
-        public int AuthorizationStrategyForDelete { get; set; }
     }
 }
